Some fun projects
BinaryTree: Adds lower numbers to the left side of the root and higher numbers to the right side.
It does this in a lower higher symmetrie.

makeWordWithLettersGiven: Checks if a word can be made by the given letters.
Let's say you have the letters dlrow ollehh and the word hello world this will return true (even though there are 2 h the word can still be made).
If you have the letters dlrow olle the method will return false because hello world can't be made. The space is also important if you want to check two words like hello world (dlrowolleh returns false because there is no space in between).

permutation: A method that returns all the permutations of a number combination.